# docs/developer-docs/public/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tari_indexer",
    "description": "The Tari indexer implementation",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.41.0"
  },
  "paths": {
...
    "/substates/fetch": {
      "post": {
        "tags": [
          "handlers::substates"
        ],
        "description": "Fetches several substates by their IDs",
        "operationId": "fetch_substates",
        "responses": {
          "200": {
            "description": "Substates details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSubstatesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Too many substates requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch substates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Indexer is still syncing, or no committee answers for the substate's shard group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/substates/{substate_id}": {
      "get": {
        "tags": [
          "handlers::substates"
        ],
        "description": "Fetches a substate by ID",
        "operationId": "get_substate",
        "parameters": [
          {
            "name": "substate_id",
            "in": "path",
            "description": "The substate ID to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "local_search_only",
            "in": "query",
            "description": "If true, only search local storage for the substate",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "version",
            "in": "query",
            "description": "Minimum version of the substate to fetch",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Substate details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSubstateResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such substate, or the version asked for has been spent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch substate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Indexer is still syncing, or no committee answers for the substate's shard group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetSubstateResponse": {
        "type": "object",
        "required": [
          "version",
          "substate",
          "verified"
        ],
        "properties": {
          "substate": {
            "type": "object"
          },
          "verified": {
            "type": "boolean",
            "description": "True when the indexer verified this substate's value against the shard group committee (via a\nmerkle proof). False when proofs are disabled, or when no committee member could supply a\nproof yet (e.g. nothing committed since an epoch change) and the value was served unverified."
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "GetSubstatesResponse": {
        "type": "object",
        "required": [
          "substates"
        ],
        "properties": {
          "substates": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}